Refactoring towards builder parameters
