+++ dbt_airflow_factory/tasks_builder/node_type.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class NodeType(Enum):
+    RUN_TEST = 1
+    MULTIPLE_DEPS_TEST = 2
+    EPHEMERAL = 3
